# Replace FNV with a faster hash function.