… src/librustc/middle/dependency_format.rs
… src/librustc/middle/privacy.rs
… src/librustc/middle/reachable.rs
… src/librustc/middle/region.rs
… src/librustc/middle/resolve_lifetime.rs
… src/librustc/middle/stability.rs
… src/librustc/session/mod.rs
… src/librustc/traits/error_reporting.rs
… src/librustc/traits/fulfill.rs
… src/librustc/traits/select.rs
… src/librustc/traits/specialize/mod.rs
… src/librustc/traits/specialize/specialization_graph.rs
… src/librustc/traits/util.rs
… src/librustc/ty/contents.rs
… src/librustc/ty/context.rs
… src/librustc/ty/fold.rs
… src/librustc/ty/mod.rs
… src/librustc/ty/trait_def.rs
… src/librustc/ty/util.rs
… src/librustc/util/nodemap.rs
… src/librustc_borrowck/borrowck/mir/elaborate_drops.rs
… src/librustc_borrowck/borrowck/mir/gather_moves.rs
… src/librustc_borrowck/borrowck/move_data.rs
… src/librustc_const_eval/_match.rs
… src/librustc_data_structures/fx.rs
… src/librustc_data_structures/obligation_forest/mod.rs
… src/librustc_data_structures/snapshot_map/mod.rs
… src/librustc_incremental/assert_dep_graph.rs
… src/librustc_incremental/calculate_svh/mod.rs
… src/librustc_incremental/persist/data.rs