@arthurprs 2016-10-17
Do we have any backing data for this algorithm? Maybe from the Firefox development process/source? Smhasher run?