@arthurprs 2016-10-17
I think it's worth discussing a couple more things while we're at it, so we nail this for good.

* do calculations considering an usize sized hash, this will help a lot in 32bit systems. It's fine to just expand it to u64 on finish. Std HashMap hashes will eventually become usize #36567 anyway.

* this works at the same byte at a time for &str/&[int] slices so the improvement is coming exclusively from integral hashing. I'm curious if we can process usize bytes at a time before falling back to the byte at a time.