@nnethercote 2016-10-17
I think you miswrote your second dot point... but I did some ad hoc profiling and found that the *vast* majority of occurrences are `write_u32` and `write_u64`. `write` was less than 1% of occurrences.