@arthurprs 2016-10-17
Did I? I can't find it. I guess I need more more coffee.

Interesting, but I'm almost sure it'll show up when we eventually move everything away from siphasher (string interner for example).