@bluss 2016-10-17
Fnv and SipHasher both have the property that the stream of bytes to hash is "untyped": a u16 fed as u16 or its byte representation is hashed the same way.

But I don't think that the `Hash` trait expects or requires that contract in any way, and that this hash function's "typed" approach is fine.

But what I do think is that a well behaved hasher must hash a slice of bytes the same way, regardless of how you split it into subslices. That means that any whole-word optimization for `Hasher::write` then needs to keep a state (which is exactly a thing that makes SipHasher a bit slow).