@arthurprs 2016-10-17
> But I don't think that the Hash trait expects or requires that contract in any way, and that this hash function's "typed" approach is fine.

Yeah, luckily the Hash trait doesn't impose any special streaming requirement.