@arthurprs 2016-10-17
I got curious so I ran smhasher on the 64bit (PR) and original 32bit hashes, I had to include 2 variants of each to be able to see how both modes of the hasher behave (integral and byte-byte...)

see gist for results: https://gist.github.com/arthurprs/5e57cd59586acd8c52dbb02b55711096

A few comments considering the code _in_ the PR.

Hashing integral types (write_...)

The quality is really bad but it's so cheap to calculate for integral types (what rustc seems to be using fnv for) that it's still a win for the combination of the workload + hashmap implementation. I'm fairly sure that the compiler sees the 0 seed and the hash boils down to a single _IMUL_ instruction.

Hashing slices (write_usize() + write())

The _write_usize(slice.len())_ will be faster and the _write()_ slower compared to fnv. So it could potentially regress those cases.

--

I think the right way forward is to have two hashers in the rustc codebase, one general purpose-ish and another for integral types.