@nnethercote 2016-10-17
@arthurps: Thank you for running these! I was about to do it myself but you've saved me the trouble.

Looking at the results... whelp, there are a lot of numbers there that I don't know how to interpret, though the "FAIL" results sound bad.

> The write_usize(slice.len()) will be faster and the write() slower compared to fnv. So it could potentially regress those cases.

Why will `write()` be slower? Because FNV does xor + mul, while the new hash does rol + xor + mul? I guess it'll be slightly slower, but the extra rol should be cheap compared to the mul?