@arthurprs 2016-10-18
> Why will write() be slower? Because FNV does xor + mul, while the new hash does rol + xor + mul? I guess it'll be slightly slower, but the extra rol should be cheap compared to the mul?
It's a 15% difference in my Intel Skylake processor, 690MB/s vs 800MB/s.