@nnethercote 2016-10-25
> But what I do think is that a well behaved hasher must hash a slice of bytes the same way, regardless of how you split it into subslices (as long as the order is the same). That means that any whole-word optimization for Hasher::write then needs to keep a state (which is exactly a thing that makes SipHasher a bit slow).

Are you sure? Where does that requirement come from? I was thinking about changing `write` so that it processes 4 or 8 bytes at a time and then does single-byte clean-up for any excess bytes at the end...