@nnethercote 2016-10-25
New version. I've made the following changes.
* `FxHasher` is now a separate type. `FnvHasher` still exists.
* I've converted all uses of `FnvHash{Map,Set}` to `FxHash{Map,Set}`. I did some profiling and found that `write` calls (i.e. variable-length hash cases) account for less than 0.1% of occurrences. Even when I weight them by their length, they account for less than 2% of all `FxHasher` operations. So I don't think treating variable-length cases differently is worthwhile.
* `FxHasher` now works with `usize` so that it will be faster on 32-bit machines.
* I remeasured and the speed-ups are basically unchanged from those in the first comment above.

r? @arthurps: what do you think?