@bluss 2016-10-25
@nnethercote I'm not sure; it's something that needs to be discussed and put into the documentation.

I think it's the logical rule by the construction of `Hash`. Imagine a chunked rope datastructure. It should have the same hash value, regardless of how it is chunked, as long as its whole string is the same. How it is chunked will determine how its data is fed to Hasher::write.