@bluss 2016-10-25
To make a concrete example, imagine `struct Rope (Vec<String>)`. The actual string value is the concatenation of the strings in the representation. `["a", "b"]` and `["ab"]` should have the same hash.