@nnethercote 2016-10-25
(New version removes the `println!` statements that I accidentally left in...)