@nnethercote 2016-10-25
> To make a concrete example, imagine struct Rope (Vec<String>). The actual string value is the concatenation of the strings in the representation. ["a", "b"] and ["ab"] should have the same hash.

It's reasonable that you would want this to work, but it's not clear to me that this behaviour is required or guaranteed, according to the docs or anything else. E.g. https://doc.rust-lang.org/std/hash/trait.Hasher.html is very terse. (But I would be happy to be proven wrong!)

Anyway, given that variable-length hashing is comparatively rare within the compiler it's not that big a deal either way.