@arthurprs 2016-10-25
Looks good to me. Somebody from the core team should weight about how to move this forward.

I wouldn't be worried about the Hasher having the "strict streaming" characteristic as the Hash type is "strongly typed" and will make the writes to hasher every time.