@aturon 2016-10-25
cc @rust-lang/compiler